Removed integer variable types
Array is still defined as an integer which does save space. However, regular variables have no benefit defined as an integer type. They have a detrimental effect - 1) each reference takes up an extra byte, 2) the space in memory is still the same as a float, 3) Vic BASIC has to convert the integer to a float and back again when doing operations on it so is a bit slower